ai-engine/requirements.txt
cookiecutter==2.7.1
Django==6.0.1
django-tailwind==4.4.2
fastapi==0.128.1
langchain-core==1.2.28
langchain-google-genai==4.2.1
langchain-openai==1.1.12
openai==2.31.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyMuPDF==1.27.2.2
python-decouple==3.8.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.32.5
requests-toolbelt==1.0.0
uvicorn==0.40.0
whitenoise==6.11.0